+++ tests/test_sockets.py
@@ -18,7 +18,7 @@ from pathlib import Path
 from socket import AddressFamily
 from ssl import SSLContext, SSLError
 from threading import Thread
-from typing import TYPE_CHECKING, Any, Literal, NoReturn, TypeVar, cast
+from typing import TYPE_CHECKING, Any, Literal, NoReturn, Protocol, TypeVar, cast
 from unittest import mock
 
 import psutil
@@ -61,11 +61,16 @@ from anyio import (
 )
 from anyio._core._eventloop import get_async_backend
 from anyio.abc import (
+    ConnectedUDPSocket,
+    ConnectedUNIXDatagramSocket,
     IPSockAddrType,
     Listener,
     SocketAttribute,
     SocketListener,
     SocketStream,
+    UDPSocket,
+    UNIXDatagramSocket,
+    UNIXSocketStream,
 )
 from anyio.lowlevel import checkpoint
 from anyio.streams.stapled import MultiListener
@@ -140,6 +145,60 @@ def check_asyncio_bug(anyio_backend_name: str, family: AnyIPAddressFamily) -> No
         policy = asyncio.get_event_loop_policy()
         if policy.__class__.__name__ == "WindowsProactorEventLoopPolicy":
             pytest.skip("Does not work due to a known bug (39148)")
+
+
+class SockFdFactoryProtocol(Protocol):
+    def __call__(
+        self,
+        family: socket.AddressFamily,
+        kind: socket.SocketKind,
+        *,
+        bound: bool = False,
+        connected: bool = False,
+    ) -> socket.socket | int: ...
+
+
+@pytest.fixture(
+    params=[pytest.param(False, id="sock"), pytest.param(True, id="fileno")]
+)
+def sock_or_fd_factory(
+    request: SubRequest, tmp_path_factory: TempPathFactory
+) -> SockFdFactoryProtocol:
+    def factory(
+        family: socket.AddressFamily,
+        kind: socket.SocketKind,
+        *,
+        bound: bool = False,
+        connected: bool = False,
+    ) -> socket.socket | int:
+        sock = socket.socket(family, kind)
+
+        if bound or connected:
+            if family in (socket.AF_INET, socket.AF_INET6):
+                local_addr: str | tuple[str, int] = ("localhost", 0)
+            else:
+                local_addr = str(tmp_path_factory.mktemp("anyio") / "socket")
+
+        if bound:
+            sock.bind(local_addr)
+            if kind == socket.SOCK_STREAM:
+                sock.listen()
+        elif connected:
+            server_sock = socket.socket(family, kind)
+            request.addfinalizer(server_sock.close)
+            server_sock.bind(local_addr)
+            if kind == socket.SOCK_STREAM:
+                server_sock.listen()
+
+            sock.connect(server_sock.getsockname())
+
+        if request.param:
+            return sock.detach()
+        else:
+            request.addfinalizer(sock.close)
+            return sock
+
+    return factory
 
 
 _T = TypeVar("_T")
@@ -573,6 +632,48 @@ class TestTCPStream:
             )
             assert not caplog_text
 
+    async def test_from_socket(
+        self, family: AnyIPAddressFamily, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(family, socket.SOCK_STREAM, connected=True)
+        async with await SocketStream.from_socket(sock_or_fd) as stream:
+            assert isinstance(stream, SocketStream)
+            assert stream.extra(SocketAttribute.family) == family
+
+    async def test_from_socket_not_a_socket(self) -> None:
+        with pytest.raises(
+            TypeError,
+            match="expected an int or socket, got str instead",
+        ):
+            await SocketStream.from_socket("foo")  # type: ignore[arg-type]
+
+    async def test_from_socket_pass_file_fd(self, tmp_path: Path) -> None:
+        with pytest.raises(
+            ValueError,
+            match="the file descriptor does not refer to a socket",
+        ):
+            with tmp_path.joinpath("foo").open("wb") as fd:
+                await SocketStream.from_socket(fd.fileno())
+
+    async def test_from_socket_wrong_socket_type(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_INET, socket.SocketKind.SOCK_DGRAM, connected=True
+        )
+        with pytest.raises(
+            ValueError,
+            match="socket type mismatch: expected SOCK_STREAM, got SOCK_DGRAM",
+        ):
+            await SocketStream.from_socket(sock_or_fd)
+
+    async def test_from_socket_not_connected(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(socket.AF_INET, socket.SOCK_STREAM, bound=True)
+        with pytest.raises(ValueError, match="the socket must be connected"):
+            await SocketStream.from_socket(sock_or_fd)
+
 
 @pytest.mark.network
 class TestTCPListener:
@@ -787,6 +888,21 @@ class TestTCPListener:
 
         async with await create_tcp_listener(local_host=link_local_ipv6_address):
             pass
+
+    async def test_from_socket(
+        self, family: AnyIPAddressFamily, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(family, socket.SOCK_STREAM, bound=True)
+        async with await SocketListener.from_socket(sock_or_fd) as listener:
+            assert isinstance(listener, SocketListener)
+            assert listener.extra(SocketAttribute.family) == family
+
+    async def test_from_socket_not_bound(
+        self, family: AnyIPAddressFamily, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(family, socket.SOCK_STREAM)
+        with pytest.raises(ValueError, match="the socket must be bound"):
+            await SocketListener.from_socket(sock_or_fd)
 
 
 @pytest.mark.skipif(
@@ -1127,6 +1243,46 @@ class TestUNIXStream:
             async with await connect_unix(actual_path):
                 pass
 
+    async def test_from_socket(
+        self, socket_path: Path, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_UNIX, socket.SOCK_STREAM, connected=True
+        )
+        async with await UNIXSocketStream.from_socket(sock_or_fd) as stream:
+            assert isinstance(stream, UNIXSocketStream)
+
+    async def test_from_socket_wrong_socket_type(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_UNIX, socket.SOCK_DGRAM, connected=True
+        )
+        with pytest.raises(
+            ValueError,
+            match="socket type mismatch: expected SOCK_STREAM, got SOCK_DGRAM",
+        ):
+            await UNIXSocketStream.from_socket(sock_or_fd)
+
+    async def test_from_socket_wrong_address_family(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_INET, socket.SOCK_STREAM, connected=True
+        )
+        with pytest.raises(
+            ValueError,
+            match="address family mismatch: expected AF_UNIX, got AF_INET",
+        ):
+            await UNIXSocketStream.from_socket(sock_or_fd)
+
+    async def test_from_socket_not_connected(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(socket.AF_UNIX, socket.SOCK_STREAM)
+        with pytest.raises(ValueError, match="the socket must be connected"):
+            await UNIXSocketStream.from_socket(sock_or_fd)
+
 
 @pytest.mark.skipif(
     sys.platform == "win32", reason="UNIX sockets are not available on Windows"
@@ -1187,7 +1343,7 @@ class TestUNIXListener:
     async def test_socket_options(self, socket_path: Path) -> None:
         async with await create_unix_listener(socket_path) as listener:
             listener_socket = listener.extra(SocketAttribute.raw_socket)
-            assert listener_socket.family == socket.AddressFamily.AF_UNIX
+            assert listener_socket.family == socket.AF_UNIX
             listener_socket.setsockopt(socket.SOL_SOCKET, socket.SO_RCVBUF, 80000)
             assert listener_socket.getsockopt(socket.SOL_SOCKET, socket.SO_RCVBUF) in (
                 80000,
@@ -1250,6 +1406,12 @@ class TestUNIXListener:
         async with await create_unix_listener(real_path):
             pass
 
+    async def test_from_socket(self, sock_or_fd_factory: SockFdFactoryProtocol) -> None:
+        sock_or_fd = sock_or_fd_factory(socket.AF_UNIX, socket.SOCK_STREAM, bound=True)
+        async with await SocketListener.from_socket(sock_or_fd) as listener:
+            assert isinstance(listener, SocketListener)
+            assert listener.extra(SocketAttribute.family) == socket.AF_UNIX
+
 
 async def test_multi_listener(tmp_path_factory: TempPathFactory) -> None:
     async def handle(stream: SocketStream) -> None:
@@ -1272,8 +1434,7 @@ async def test_multi_listener(tmp_path_factory: TempPathFactory) -> None:
                     local_address = listener.extra(SocketAttribute.local_address)
                     if (
                         sys.platform != "win32"
-                        and listener.extra(SocketAttribute.family)
-                        == socket.AddressFamily.AF_UNIX
+                        and listener.extra(SocketAttribute.family) == socket.AF_UNIX
                     ):
                         assert isinstance(local_address, str)
                         stream: SocketStream = await connect_unix(local_address)
@@ -1417,6 +1578,26 @@ class TestUDPSocket:
             )
             assert local_address[1] > 0
 
+    async def test_from_socket(
+        self, family: AnyIPAddressFamily, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(family, socket.SOCK_DGRAM, bound=True)
+        async with await UDPSocket.from_socket(sock_or_fd) as udp_socket:
+            assert isinstance(udp_socket, UDPSocket)
+            assert udp_socket.extra(SocketAttribute.family) == family
+
+    async def test_from_socket_wrong_socket_type(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_INET, socket.SOCK_STREAM, connected=True
+        )
+        with pytest.raises(
+            ValueError,
+            match="socket type mismatch: expected SOCK_DGRAM, got SOCK_STREAM",
+        ):
+            await UDPSocket.from_socket(sock_or_fd)
+
 
 @pytest.mark.network
 @pytest.mark.usefixtures("check_asyncio_bug")
@@ -1539,6 +1720,33 @@ class TestConnectedUDPSocket:
         await udp.aclose()
         with pytest.raises(ClosedResourceError):
             await udp.send(b"foo")
+
+    async def test_from_socket(
+        self, family: AnyIPAddressFamily, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(family, socket.SOCK_DGRAM, connected=True)
+        async with await ConnectedUDPSocket.from_socket(sock_or_fd) as udp_socket:
+            assert isinstance(udp_socket, ConnectedUDPSocket)
+            assert udp_socket.extra(SocketAttribute.family) == family
+
+    async def test_from_socket_wrong_socket_type(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_INET, socket.SOCK_STREAM, connected=True
+        )
+        with pytest.raises(
+            ValueError,
+            match="socket type mismatch: expected SOCK_DGRAM, got SOCK_STREAM",
+        ):
+            await ConnectedUDPSocket.from_socket(sock_or_fd)
+
+    async def test_from_socket_not_connected(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(socket.AF_INET, socket.SOCK_DGRAM, bound=True)
+        with pytest.raises(ValueError, match="the socket must be connected"):
+            await ConnectedUDPSocket.from_socket(sock_or_fd)
 
 
 @pytest.mark.skipif(
@@ -1681,6 +1889,22 @@ class TestUNIXDatagramSocket:
         real_path = str(socket_path).encode() + b"\xf0"
         async with await create_unix_datagram_socket(local_path=real_path):
             pass
+
+    async def test_from_socket(self, sock_or_fd_factory: SockFdFactoryProtocol) -> None:
+        sock_or_fd = sock_or_fd_factory(socket.AF_UNIX, socket.SOCK_DGRAM)
+        async with await UNIXDatagramSocket.from_socket(sock_or_fd) as udp_socket:
+            assert isinstance(udp_socket, UNIXDatagramSocket)
+            assert udp_socket.extra(SocketAttribute.family) == socket.AF_UNIX
+
+    async def test_from_socket_wrong_socket_type(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(socket.AF_UNIX, socket.SOCK_STREAM)
+        with pytest.raises(
+            ValueError,
+            match="socket type mismatch: expected SOCK_DGRAM, got SOCK_STREAM",
+        ):
+            await UNIXDatagramSocket.from_socket(sock_or_fd)
 
 
 @pytest.mark.skipif(
@@ -1856,6 +2080,35 @@ class TestConnectedUNIXDatagramSocket:
         with pytest.raises(ClosedResourceError):
             await udp.send(b"foo")
 
+    async def test_from_socket(self, sock_or_fd_factory: SockFdFactoryProtocol) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_UNIX, socket.SOCK_DGRAM, connected=True
+        )
+        async with await ConnectedUNIXDatagramSocket.from_socket(
+            sock_or_fd
+        ) as udp_socket:
+            assert isinstance(udp_socket, ConnectedUNIXDatagramSocket)
+            assert udp_socket.extra(SocketAttribute.family) == socket.AF_UNIX
+
+    async def test_from_socket_wrong_socket_type(
+        self, socket_path: Path, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(
+            socket.AF_UNIX, socket.SOCK_STREAM, connected=True
+        )
+        with pytest.raises(
+            ValueError,
+            match="socket type mismatch: expected SOCK_DGRAM, got SOCK_STREAM",
+        ):
+            await ConnectedUNIXDatagramSocket.from_socket(sock_or_fd)
+
+    async def test_from_socket_not_connected(
+        self, sock_or_fd_factory: SockFdFactoryProtocol
+    ) -> None:
+        sock_or_fd = sock_or_fd_factory(socket.AF_UNIX, socket.SOCK_DGRAM, bound=True)
+        with pytest.raises(ValueError, match="the socket must be connected"):
+            await ConnectedUNIXDatagramSocket.from_socket(sock_or_fd)
+
 
 async def test_tcp_connectable(mocker: MockerFixture) -> None:
     mock_connect_tcp = mocker.patch("anyio._core._sockets.connect_tcp")
@@ -1921,7 +2174,7 @@ class TestAsConnectable:
 
     def test_tls_explicit_context(self, client_context: SSLContext) -> None:
         connectable = as_connectable(
-            "/foo/bar", tls=client_context, tls_hostname="example.com"
+            "/foo/bar", tls=True, ssl_context=client_context, tls_hostname="example.com"
         )
         assert isinstance(connectable, TLSConnectable)
         assert connectable.ssl_context is client_context
@@ -1941,9 +2194,7 @@ async def test_getaddrinfo() -> None:
     assert correct != wrong
 
 
-@pytest.mark.parametrize(
-    "sock_type", [socket.SOCK_STREAM, socket.SocketKind.SOCK_STREAM]
-)
+@pytest.mark.parametrize("sock_type", [socket.SOCK_STREAM, socket.SOCK_STREAM])
 async def test_getaddrinfo_ipv6addr(
     sock_type: Literal[socket.SocketKind.SOCK_STREAM],
 ) -> None:
@@ -1951,8 +2202,8 @@ async def test_getaddrinfo_ipv6addr(
     proto = 0 if platform.system() == "Windows" else 6
     assert await getaddrinfo("::1", 0, type=sock_type) == [
         (
-            socket.AddressFamily.AF_INET6,
-            socket.SocketKind.SOCK_STREAM,
+            socket.AF_INET6,
+            socket.SOCK_STREAM,
             proto,
             "",
             ("::1", 0),
@@ -1961,9 +2212,7 @@ async def test_getaddrinfo_ipv6addr(
 
 
 async def test_getaddrinfo_ipv6_disabled() -> None:
-    gai_result = [
-        (AddressFamily.AF_INET6, socket.SocketKind.SOCK_STREAM, 6, "", (1, b""))
-    ]
+    gai_result = [(AddressFamily.AF_INET6, socket.SOCK_STREAM, 6, "", (1, b""))]
     with mock.patch.object(get_async_backend(), "getaddrinfo", return_value=gai_result):
         assert await getaddrinfo("::1", 0) == []